refactor: 각 기물 점수 구현 및 Soldier 추상 클래스로 리팩터링
