+++ src/main/java/janggi/domain/Dynasty.java
@@ -1,5 +1,11 @@
 package janggi.domain;
 
 public enum Dynasty {
-    CHU, HAN, EMPTY
+    CHU(72), HAN(73.5), EMPTY(0);
+
+    private final double initialScore;
+
+    Dynasty(double initialScore) {
+        this.initialScore = initialScore;
+    }
 }
