+++ src/main/java/janggi/domain/piece/General.java
@@ -25,6 +25,11 @@ public class General extends Piece {
     }
 
     @Override
+    public int score() {
+        return 0;
+    }
+
+    @Override
     public boolean isSameType(Piece piece) {
         return piece instanceof General;
     }
